+++ changelog.md
@@ -15,6 +15,7 @@ Last change: 21-JUN-2026 CEST
 - Remove a ROI by double-clicking it in the preview
 - Loading a movie with corrupted metadata lets the user specify the most important ones without errors
 - 3D calibration allows for multiple FOVs per z position (thanks to Aditya Ajay Chhatre for the suggestion)
+- Slider at the bottom of the app was added to allow easy navigation across frames
 - New keyboard shortcuts for navigating movies (move by 10/100/1,000 frames)
 - Slight adjustments to some status bar messages
 - Fixed a gap of roughly one box size in the identified spots along the borders between adjacent (e.g. overlapping) ROIs